+++ qom/solvers/ODESolver.py
@@ -6,7 +6,7 @@
 __name__    = 'qom.solvers.ODESolver'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2021-01-04'
-__updated__ = '2021-01-12'
+__updated__ = '2021-01-13'
 
 # dependencies
 from typing import Union
@@ -19,6 +19,8 @@ logger = logging.getLogger(__name__)
 
 # datatypes
 t_array = Union[list, np.matrix, np.ndarray]
+
+# TODO: Add old API submodules in `set_integrator_params`.
 
 class ODESolver():
     r"""Class to handle ODE solver.
@@ -35,13 +37,26 @@ class ODESolver():
         Initial values for the function.
     c : list or np.matrix or np.ndarray, optional
         Constants for the function.
-    method : str
+    method : str, optional
         Method used to solve the ODEs:
             'BDF': :class:`scipy.integrate.BDF`.
             'DOP853': :class:`scipy.integrate.DOP853`.
+            'LSODA': :class:`scipy.integrate.LSODA`.
             'ode': :class:`scipy.integrate.ode`.
+            'Radau': :class:`scipy.integrate.Radau`.
+            'RK23': :class:`scipy.integrate.RK23`.
             'RK45': :class:`scipy.integrate.RK45` (default).
     """
+
+    # attributes
+    new_APIs = {
+        'BDF': si.BDF,
+        'DOP853': si.DOP853,
+        'LSODA': si.LSODA,
+        'Radau': si.Radau,
+        'RK23': si.RK23,
+        'RK45': si.RK45,
+    }
 
     @property
     def c(self):
@@ -88,7 +103,10 @@ class ODESolver():
         """str: Method used to solve the ODEs:
             'BDF': :class:`scipy.integrate.BDF`.
             'DOP853': :class:`scipy.integrate.DOP853`.
+            'LSODA': :class:`scipy.integrate.LSODA`.
             'ode': :class:`scipy.integrate.ode`.
+            'Radau': :class:`scipy.integrate.Radau`.
+            'RK23': :class:`scipy.integrate.RK23`.
             'RK45': :class:`scipy.integrate.RK45` (default)."""
 
         return self.__method
@@ -148,30 +166,27 @@ class ODESolver():
 
         # extract frequently used variables
         solver_type = self.params.get('type', solver_type)
-        _new_APIs = {
-            'BDF': si.BDF,
-            'DOP853': si.DOP853,
-            'RK45': si.RK45,
-        }
 
-        # FORTRAN-based old API method of scipy.integrate
-        if self.method not in _new_APIs:
-            # update method name
-            self.method = 'ode'
+        # check method
+        if self.method not in self.new_APIs:
+            # FORTRAN-based old API method of scipy.integrate
+            if self.method == 'ode':
+                # get integrator
+                self.integrator = si.ode(self.func)
 
-            # get integrator
-            self.integrator = si.ode(self.func)
+                # for complex ode solver
+                if solver_type.find('complex') != -1:
+                    self.integrator.set_integrator('zvode')
+                    
+                # set initial values and constants
+                self.integrator.set_initial_value(self.iv, T[0])
 
-            # for complex ode solver
-            if solver_type.find('complex') != -1:
-                self.integrator.set_integrator('zvode')
-                
-            # set initial values and constants
-            self.integrator.set_initial_value(self.iv, T[0])
-
-            # set constants
-            if self.c is not None:
-                self.integrator.set_f_params(self.c)
+                # set constants
+                if self.c is not None:
+                    self.integrator.set_f_params(self.c)
+            # default to RK45
+            else:
+                self.method = 'RK45'
 
     def step(self, t_e, t_s=0):
         """Method to perform one step of the integration.
@@ -189,11 +204,8 @@ class ODESolver():
             Values of the integrated variables at the end time.
         """
 
-        # extract frequently used variables
-        _new_APIs = ['BDF', 'DOP853', 'RK45']
-
         # new API methods
-        if self.method in _new_APIs:
+        if self.method in self.new_APIs:
             # solve
             _sols = si.solve_ivp(self.func, (t_s, t_e), self.iv, method=self.method, args=(self.c, ))